requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-multipart==0.0.18
sqlalchemy==1.4.54
psycopg2-binary==2.9.1
python-jose==3.3.0
passlib==1.7.4
opencv-python-headless==4.8.1.78
pytesseract==0.3.10
pdf2image==1.16.3
pillow==10.0.0
numpy==1.24.3
PyPDF2==3.0.1
reportlab==4.0.4
PyMuPDF==1.23.7
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
python-dotenv==1.0.0
tqdm==4.66.1
pydantic-settings==2.6.1